+++ src/main/java/com/example/rentalhousemanagementsystem/repository/PaymentRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface PaymentRepository extends JpaRepository<Payment,Long> {
-    boolean existsByPaymentId(long uid);
+public interface PaymentRepository extends JpaRepository<Payment,Integer> {
+    boolean existsByPaymentId(int uid);
 
     @Modifying
     @Query(value = "UPDATE PAYMENT SET status =?2 WHERE payment_id=?1", nativeQuery = true)
